add support for notification message when member is removed or added
